fix(infra-test): use tsx to run test.ts files
