+++ .github/workflows/ci.yml
@@ -18,7 +18,7 @@ jobs:
       - name: Setup Node
         uses: actions/setup-node@v3
         with:
-          node-version: 16
+          node-version: 18
       - name: Install sg
         run: npm install -g @ast-grep/cli
       - name: Install deps
