+++ fixture/nest/test.py
@@ -1,1 +1,3 @@
-print("no-print warning should be displayed here")
+print("no-print warning should be displayed here 1")
+print("no-print warning should be displayed here 2")
+print("no-print warning should be displayed here 3")
