+++ scripts/runTest.mjs
@@ -1,0 +1,30 @@
+import path from 'node:path'
+import { fileURLToPath } from 'node:url'
+
+import { runTests } from '@vscode/test-electron'
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url))
+
+async function main() {
+  try {
+    // The folder containing the Extension Manifest package.json
+    // Passed to `--extensionDevelopmentPath`
+    const extensionDevelopmentPath = path.resolve(__dirname, '../')
+
+    // The path to test runner
+    // Passed to --extensionTestsPath
+    const extensionTestsPath = path.resolve(__dirname, './integration-test.cjs')
+
+    // Download VS Code, unzip it and run the integration test
+    await runTests({
+      extensionDevelopmentPath,
+      extensionTestsPath,
+      launchArgs: ['--disable-extensions']
+    })
+  } catch (err) {
+    console.error('Failed to run tests', err)
+    process.exit(1)
+  }
+}
+
+main()
